adding group rules

+++ 1index.php
@@ -39,6 +39,7 @@
                     </thead>
                     <tbody>
                     <?php foreach($tasks as $item): ?>
+                        <?php if(checkRule()>5) ?>
                         <tr>
                         <td><?=$item["assigner"]?></td>
                         <td><?=$item["assigner_email"]?></td>
@@ -49,10 +50,13 @@
                         <td><time datetime="<?=$item["beginning"]?>"><?=date("j F Y \г\. \в H:i", strtotime($item["beginning"]))?></time></td>
                         <td><time datetime="<?=$item["deadline"]?>"><?=date("j F Y \г\. \в H:i", strtotime($item["deadline"]))?></time> </td>
                         <td>
+                            
+                        <?php if(checkRule() >= 60): ?>
                             <a href="update_task.php?id=<?=$item["id_task"]?>">Изменить/</a>
                             <a href="delete_task.php?id=<?=$item["id_task"]?>">Удалить</a>
                         
                     </td>
+                    <?php endif; ?>
                     </tr>
                     <?php endforeach; ?>
                     </tbody>
